+++ firebase-vscode/webviews/globals/ux-text.ts
@@ -1,11 +1,4 @@
 export const TEXT = {
-  INIT_HOSTING_BUTTON: "Host your Web App",
-
-  INIT_HOSTING_DESCRIPTION:
-    "Deploy your app with Firebase Hosting" +
-    ", a high-performance static web host backed by a global CDN",
-
-  INIT_HOSTING_PROGRESS: "Initializing...",
 
   LOGIN_PROGRESS: "Checking login",
 
@@ -18,6 +11,8 @@ export const TEXT = {
   VSCE_SERVICE_ACCOUNT_SELECTION_ITEM: "Service account",
 
   MONOSPACE_LOGIN_FAIL: "Unable to find default credentials",
+
+  LOGIN_IN_PROGRESS: "Login in progress",
 
   GOOGLE_SIGN_IN: "Sign in with Google",