+++ src/emulator/downloadableEmulators.ts
@@ -57,20 +57,20 @@ const EMULATOR_UPDATE_DETAILS: { [s in DownloadableEmulators]: EmulatorUpdateDet
   dataconnect:
     process.platform === "darwin"
       ? {
-          version: "1.2.0",
-          expectedSize: 23954240,
-          expectedChecksum: "0f250761959519bb5a28fed76ceab2cb",
+          version: "1.2.4",
+          expectedSize: 24097600,
+          expectedChecksum: "e0a344620b71d64b79d99b2c358e7646",
         }
       : process.platform === "win32"
         ? {
-            version: "1.2.0",
-            expectedSize: 24360960,
-            expectedChecksum: "168ce32c742e1d26037c52bdbb7d871c",
+            version: "1.2.4",
+            expectedSize: 24509440,
+            expectedChecksum: "44ebc07e481a85bd735fe8007fab3efe",
           }
         : {
-            version: "1.2.0",
-            expectedSize: 23970052,
-            expectedChecksum: "2ca17e4009a9ebae0f7c983bafff2ee6",
+            version: "1.2.4",
+            expectedSize: 24010904,
+            expectedChecksum: "bf10e866f4c4e132bf8115460508e6e2",
           },
 };
 
@@ -292,11 +292,8 @@ const Commands: { [s in DownloadableEmulators]: DownloadableEmulatorCommand } = 
     optionalArgs: [
       "listen",
       "config_dir",
-      "project_id",
-      "service_location",
       "disable_sdk_generation",
       "resolvers_emulator",
-      "vertex_location",
       "rpc_retry_count",
     ],
     joinArgs: true,
@@ -304,7 +301,7 @@ const Commands: { [s in DownloadableEmulators]: DownloadableEmulatorCommand } = 
   },
 };
 
-export function getExecPath(name: DownloadableEmulators): string {
+function getExecPath(name: DownloadableEmulators): string {
   const details = getDownloadDetails(name);
   return details.binaryPath || details.downloadPath;
 }