+++ CHANGELOG.md
@@ -1,3 +1,1 @@
-- Fixes issue where storage emulator throws an error due to non-standard whitespaces in filenames (#6834).
-- Fixes issue where some emulators would fail to start when their path contained a whitespace (#7313)
-- Adds prompt for Postgres connection string to `setup:emulators:dataconnect`.
+- `firebase init dataconnect` now can pull down deployed GQL files.